+++ next.config.mjs
@@ -10,6 +10,7 @@
 const nextConfig = {
   reactStrictMode: true,
   poweredByHeader: false,
+  devIndicators: false,
 };
 
 export default nextConfig;